\name{lvnetCompare}
\alias{lvnetCompare}
\alias{anova.lvnet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compare lvnet objects
}
\description{
Compares several results of \code{\link{lvnet}}
}
\usage{
lvnetCompare(...)
\method{anova}{lvnet}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{object}{An \code{lvnet} object}
  \item{\dots}{
Any number of \code{lvnet} objects. Arguments can be named to make the resulting table named.
}
}
\author{
Sacha Epskamp <mail@sachaepskamp.com>
}

\seealso{
\code{\link{lvnet}}
}
